+++ lib/discourse_assign/engine.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module ::DiscourseAssign
+  class Engine < ::Rails::Engine
+    engine_name PLUGIN_NAME
+    isolate_namespace DiscourseAssign
+  end
+end
